LayoutTests/fast/css/text-wrap-pretty-disabled.html

<!-- webkit-test-runner [ CSSTextWrapPrettyEnabled=false ]-->
<!DOCTYPE html>
<html>
<head>
<title>Test text-wrap parsing when text-wrap: pretty is disabled</title>
<meta charset="utf-8">
</head>
<body>
<p>When text-wrap: pretty is disabled, the following should not parse:</p>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../imported/w3c/web-platform-tests/css/support/parsing-testcommon.js"></script>
<script>
test_invalid_value("text-wrap-style", "pretty");
test_invalid_value("text-wrap", "pretty");
test_invalid_value("text-wrap", "pretty wrap");
test_invalid_value("text-wrap", "pretty nowrap");
test_invalid_value("text-wrap", "wrap pretty");
test_invalid_value("text-wrap", "nowrap pretty");
</script>
</body>
</html>
